refactor : Design System 의존성 변경
