+++ aljo/Projects/Features/Auth/Project.swift
@@ -2,13 +2,23 @@ import ProjectDescription
 import ProjectDescriptionHelpers
 import MyPlugin
 
-
 let targets: [Target] = [
   Target(
-    name: "DesignSystem",
+    name: "Auth",
     platform: .iOS,
     product: .framework,
-    bundleId: "com.design.designSystem",
+    bundleId: "com.asap.feature.auth",
+    deploymentTarget: .iOS(targetVersion: "14.0", devices: .iphone),
+    sources: ["Sources/**"],
+    dependencies: [
+      .project(target: "BaseFeature", path: "../BaseFeature")
+    ]
+  ),
+  Target(
+    name: "AuthApp",
+    platform: .iOS,
+    product: .app,
+    bundleId: "com.asap.feature.authapp",
     deploymentTarget: .iOS(targetVersion: "14.0", devices: .iphone),
     sources: ["Sources/**"]
   )
@@ -18,4 +28,4 @@ let targets: [Target] = [
 let localHelper = LocalHelper(name: "MyPlugin")
 
 // Creates our project using a helper function defined in ProjectDescriptionHelpers
-let project = Project(name: "DesignSystem", targets: targets)
+let project = Project(name: "Auth", targets: targets)
