+++ aljo/Projects/Features/BaseFeature/Sources/BaseFeatureImpl.swift
@@ -1,0 +1,6 @@
+import UIKit
+import AJUIKit
+
+protocol BaseViewControllable: UIViewController {
+  
+}
