+++ aljo/Projects/Features/BaseFeature/Tuist/Config.swift
@@ -1,0 +1,7 @@
+import ProjectDescription
+
+let config = Config(
+    plugins: [
+        .local(path: .relativeToManifest("../../Plugins/BaseFeature")),
+    ]
+)
