--- aljo/Projects/Features/Sources/MyPage/Plugins/MyPage/Package.swift
@@ -1,15 +1,0 @@
-// swift-tools-version: 5.4
-
-import PackageDescription
-
-let package = Package(
-    name: "MyPlugin",
-    products: [
-        .executable(name: "tuist-my-cli", targets: ["tuist-my-cli"]),
-    ],
-    targets: [
-        .executableTarget(
-            name: "tuist-my-cli"
-        ),
-    ]
-)
