--- aljo/Projects/Features/Sources/MyPage/Plugins/MyPage/Plugin.swift
@@ -1,3 +1,0 @@
-import ProjectDescription
-
-let plugin = Plugin(name: "MyPlugin")
