--- aljo/Projects/Features/Sources/MyPage/Plugins/MyPage/ProjectDescriptionHelpers/LocalHelper.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-public struct LocalHelper {
-    let name: String
-
-    public init(name: String) {
-        self.name = name
-    }
-}
