--- aljo/Projects/Features/Sources/MyPage/Plugins/MyPage/Sources/tuist-my-cli/main.swift
@@ -1,1 +1,0 @@
-print("Hello, from your Tuist Task")
