--- aljo/Projects/Features/Tuist/Config.swift
@@ -1,7 +1,0 @@
-import ProjectDescription
-
-let config = Config(
-    plugins: [
-        .local(path: .relativeToManifest("../../Plugins/Features")),
-    ]
-)
